+++ ChangeLog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- `wigner_3j`  now _always_ returns a contiguous copy regardless of dtype or device
 
 ## [0.4.4] - 2021-12-15
 ### Fixed